refactored decoding of 'iovl' images
