+++ libheif/codecs/mask_image.h
@@ -91,6 +91,8 @@ public:
 
   const heif_color_profile_nclx* get_forced_output_nclx() const override { return nullptr; }
 
+  heif_compression_format get_compression_format() const override { return heif_compression_mask; }
+
   bool is_ispe_essential() const override { return true; }
 
   Result<CodedImageData> encode(const std::shared_ptr<HeifPixelImage>& image,
